refactor(pruner): unexport PruneUpto and OldestRetainedBlock
These functions were exported only so the external pruner_test package
could reach them; neither is part of the pruner package's cross-package
API. Unexport them to `pruneUpto` / `oldestRetainedBlock` and relocate
their direct tests into internal (package pruner) test files, following
the existing accessors_internal_test.go precedent:

  - pruneUpto lifecycle tests -> pruneupto_internal_test.go
  - oldestRetainedBlock tests -> retention_internal_test.go
  - running_event_filter_test.go moves to package pruner since it uses
    pruneUpto to set up pruned state before exercising the exported
    InitializeRunningEventFilter

TestPruneBlockDataUpto stays in the external test package: it covers the
still-exported PruneBlockDataUpto.

FindOldestBlockAtOrAfter (also named in #3671) is intentionally left
exported: migration/historyprunner consumes it (and ErrNoBlockInWindow)
as production API, so it is not test-only.

Closes #3671
