+++ pruner/accessors_test.go
@@ -1,7 +1,6 @@
 package pruner_test
 
 import (
-	"context"
 	"testing"
 
 	"github.com/NethermindEth/juno/core"
@@ -10,185 +9,11 @@ import (
 	"github.com/NethermindEth/juno/pruner"
 	"github.com/NethermindEth/juno/pruner/testutils"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
-
-const testTargetBatchByteSize = 96 * db.Megabyte
-
-// TestPruneUpto verifies the complete prune lifecycle with realistic block
-// counts: what survives, what gets deleted, and the two intentional
-// carve-outs (BlockHashLag headers and the single hash→number mapping at
-// endExclusive-1).
-//
-// Setup: 30 blocks (0..29). BlockHashLag = 10. Call PruneUpto(20).
-//
-// Expected after the call:
-//
-//	Blocks [0, 10):  fully pruned (below the lag floor).
-//	Blocks [10, 20): commitments / state updates / txs / tx-hash lookups /
-//	                 state history pruned. Headers KEPT (lag carve-out).
-//	                 Hash→number mapping KEPT only for block 19; pruned
-//	                 for blocks 10..18.
-//	Blocks [20, 30): untouched.
-func TestPruneUpto(t *testing.T) {
-	const totalBlocks uint64 = 30
-	const endExclusive uint64 = 20
-	lag := core.BlockHashLag
-	require.GreaterOrEqual(t, endExclusive, lag)
-
-	database := testutils.NewPebbleTestDB(t)
-	blocks := make([]*testutils.StoredBlock, totalBlocks)
-	for i := range totalBlocks {
-		blocks[i] = testutils.StoreBlock(t, database, i)
-	}
-	pruned, oldestKept, err := pruner.PruneUpto(
-		t.Context(),
-		database,
-		endExclusive,
-		testTargetBatchByteSize,
-	)
-	require.NoError(t, err)
-	assert.Equal(t, endExclusive, pruned)
-	assert.Equal(t, endExclusive, oldestKept)
-
-	testutils.AssertPostPruneState(t, database, blocks, endExclusive, lag)
-}
-
-func TestPruneUpto_NoOp(t *testing.T) {
-	t.Run("empty database returns 0", func(t *testing.T) {
-		database := testutils.NewPebbleTestDB(t)
-		pruned, oldestKept, err := pruner.PruneUpto(t.Context(), database, 20, testTargetBatchByteSize)
-		require.NoError(t, err)
-		assert.Zero(t, pruned)
-		assert.Zero(t, oldestKept)
-	})
-
-	cases := []struct {
-		name           string
-		chainStart     uint64
-		endExclusive   uint64
-		wantOldestKept uint64
-	}{
-		{"endExclusive at oldestKept", 5, 5, 5},
-		{"endExclusive below oldestKept", 10, 5, 10},
-	}
-	for _, tc := range cases {
-		t.Run(tc.name+" is a no-op", func(t *testing.T) {
-			database := testutils.NewPebbleTestDB(t)
-			blocks := make([]*testutils.StoredBlock, 30)
-			for i := tc.chainStart; i < 30; i++ {
-				blocks[i] = testutils.StoreBlock(t, database, i)
-			}
-			pruned, oldestKept, err := pruner.PruneUpto(
-				t.Context(),
-				database,
-				tc.endExclusive,
-				testTargetBatchByteSize,
-			)
-			require.NoError(t, err)
-			assert.Zero(t, pruned)
-			assert.Equal(t, tc.wantOldestKept, oldestKept)
-			for i := tc.chainStart; i < 30; i++ {
-				testutils.AssertBlockExists(t, database, blocks[i])
-			}
-		})
-	}
-}
-
-// cancelAfterGets wraps a KeyValueStore and cancels the supplied context
-// after the n-th Get call. pruneHashKeyedUpto issues exactly one Get per
-// loop iteration (core.GetStateUpdateByBlockNum), so this lets a test
-// stop the sweep at a precise block boundary.
-type cancelAfterGets struct {
-	db.KeyValueStore
-	cancel context.CancelFunc
-	after  int
-	count  int
-}
-
-func (c *cancelAfterGets) Get(key []byte, cb func([]byte) error) error {
-	c.count++
-	if c.count == c.after {
-		c.cancel()
-	}
-	return c.KeyValueStore.Get(key, cb)
-}
-
-// TestPruneUpto_ResumeAfterMidwayCancel verifies that cancelling PruneUpto
-// after some blocks have been processed and re-running picks up exactly
-// where the first call stopped — no block is processed twice, none is
-// skipped. Asserted by summing the per-call blocksPruned counts.
-func TestPruneUpto_ResumeAfterMidwayCancel(t *testing.T) {
-	const totalBlocks uint64 = 30
-	const endExclusive uint64 = 25
-	const cancelAfter = 10 // cancel after the 10th state-update Get
-
-	database := testutils.NewPebbleTestDB(t)
-	blocks := make([]*testutils.StoredBlock, totalBlocks)
-	for i := range totalBlocks {
-		blocks[i] = testutils.StoreBlock(t, database, i)
-	}
-
-	ctx, cancel := context.WithCancel(t.Context())
-	wrapped := &cancelAfterGets{KeyValueStore: database, cancel: cancel, after: cancelAfter}
-
-	pruned1, oldestKept1, err := pruner.PruneUpto(ctx, wrapped, endExclusive, testTargetBatchByteSize)
-	require.NoError(t, err)
-	require.Greater(t, pruned1, uint64(0), "first call should make progress before cancel")
-	require.Less(t, pruned1, endExclusive, "first call must not finish the full window")
-	assert.Equal(t, pruned1, oldestKept1, "oldestKept after partial prune == pruned (start was 0)")
-
-	pruned2, oldestKept2, err := pruner.PruneUpto(
-		t.Context(), database, endExclusive, testTargetBatchByteSize,
-	)
-	require.NoError(t, err)
-	assert.Equal(t, endExclusive, oldestKept2)
-
-	assert.Equal(t, endExclusive, pruned1+pruned2,
-		"sum of blocks pruned across both calls must equal the full window — "+
-			"any duplicate or skipped block would break this invariant")
-
-	// End-state: blocks below the lag floor are gone, blocks ≥ endExclusive intact.
-	for i := range endExclusive - core.BlockHashLag {
-		testutils.AssertBlockPruned(t, database, blocks[i])
-	}
-	for i := endExclusive; i < totalBlocks; i++ {
-		testutils.AssertBlockExists(t, database, blocks[i])
-	}
-}
-
-// TestPruneUpto_RollingCarveOut verifies that across two consecutive
-// PruneUpto calls, the hash→number carve-out rolls forward correctly:
-// after each call exactly one extra mapping survives below oldestKept,
-// and the previous extra is cleaned up by the next call's start-1 sweep.
-func TestPruneUpto_RollingCarveOut(t *testing.T) {
-	const totalBlocks uint64 = 40
-	lag := core.BlockHashLag
-
-	database := testutils.NewPebbleTestDB(t)
-	blocks := make([]*testutils.StoredBlock, totalBlocks)
-	for i := range totalBlocks {
-		blocks[i] = testutils.StoreBlock(t, database, i)
-	}
-
-	// Call 1: prune up to 20.
-	pruned, oldestKept, err := pruner.PruneUpto(t.Context(), database, 20, testTargetBatchByteSize)
-	require.NoError(t, err)
-	assert.Equal(t, uint64(20), pruned)
-	assert.Equal(t, uint64(20), oldestKept)
-	testutils.AssertPostPruneState(t, database, blocks, 20, lag)
-
-	// Call 2: prune up to 30. OldestRetainedBlock returns 20 now, so we prune 10 more.
-	pruned, oldestKept, err = pruner.PruneUpto(t.Context(), database, 30, testTargetBatchByteSize)
-	require.NoError(t, err)
-	assert.Equal(t, uint64(10), pruned)
-	assert.Equal(t, uint64(30), oldestKept)
-	testutils.AssertPostPruneState(t, database, blocks, 30, lag)
-}
 
 // TestPruneBlockDataUpto exercises the range-tombstone half on its own,
 // confirming the BlockHashLag carve-out for headers without the per-block
-// hash-keyed work that PruneUpto wraps around it.
+// hash-keyed work that pruneUpto wraps around it.
 func TestPruneBlockDataUpto(t *testing.T) {
 	const totalBlocks uint64 = 30
 	const endExclusive uint64 = 20
